requirements.txt
mysql_connector_repackaged==0.3.1
numpy==1.23.4
pandas==1.5.1
scikit_learn==1.2.1
streamlit==1.13.0
